+++ CHANGELOG.md
@@ -34,7 +34,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 ### Features
 
 * [53](https://github.com/JulianToledano/goingecko/pull/53) Add implementation for coingecko [websockets api](https://docs.coingecko.com/websocket)
-
+* [54](https://github.com/JulianToledano/goingecko/pull/54) Add coins paid endpoints
 
 ## [v3.1.0](https://github.com/JulianToledano/goingecko/releases/tag/v3.1.0) - 2025-11-28